rename expando -> collapsible
