+++ frontend/components/collapsible/index.ts
